refactor: restructure project directory
