+++ internal/terminalplot/livedashboard.go
