+++ internal/terminalplot/virtualmem.go
@@ -3,7 +3,7 @@ package terminalplot
 import (
 	"fmt"
 	"log"
-	"dcm/memory"
+	"dcm/internal/memory"
 	ui "github.com/gizak/termui/v3"
 	"github.com/gizak/termui/v3/widgets"
 )
